+++ src/shared/infrastructure/http/app.ts
@@ -327,7 +327,6 @@ export const createExpressApp = (
       controllers.getCollectionsController,
       controllers.getCollectionsForUrlController,
       controllers.getRecommendedCollectionsForUrlController,
-      controllers.getRecommendedOpenCollectionsForUrlController,
       controllers.searchCollectionsController,
       controllers.getOpenCollectionsWithContributorController,
       controllers.getCollectionFollowersController,